# Optimize vMF sampling