--- a/src/samplers/vonmisesfisher.jl
+++ b/src/samplers/vonmisesfisher.jl
@@ -6,7 +6,7 @@ struct VonMisesFisherSampler
     b::Float64
     x0::Float64
     c::Float64
-    Q::Matrix{Float64}
+    v::Vector{Float64}
 end
 
 function VonMisesFisherSampler(μ::Vector{Float64}, κ::Float64)
@@ -14,8 +14,8 @@ function VonMisesFisherSampler(μ::Vector{Float64}, κ::Float64)
     b = _vmf_bval(p, κ)
     x0 = (1.0 - b) / (1.0 + b)
     c = κ * x0 + (p - 1) * log1p(-abs2(x0))
-    Q = _vmf_rotmat(μ)
-    VonMisesFisherSampler(p, κ, b, x0, c, Q)
+    v = _vmf_householder_vec(μ)
+    VonMisesFisherSampler(p, κ, b, x0, c, v)
 end
 
 function _rand!(rng::AbstractRNG, spl::VonMisesFisherSampler,
@@ -36,7 +36,9 @@ function _rand!(rng::AbstractRNG, spl::VonMisesFisherSampler,
     end
 
     # rotate
-    mul!(x, spl.Q, t)
+    scale = 2.0 * (spl.v' * t)
+    copyto!(x, t)
+    @. x -= (scale * spl.v)
     return x
 end
 
@@ -59,8 +61,14 @@ _vmf_bval(p::Int, κ::Real) = (p - 1) / (2.0κ + sqrt(4 * abs2(κ) + abs2(p - 1)
 function _vmf_genw(rng::AbstractRNG, p, b, x0, c, κ)
     # generate the W value -- the key step in simulating vMF
     #
-    #   following movMF's document
+    #   following movMF's document for the p > 3 case
+    #   and Wenzel Jakob's document for the p == 3 case
     #
+    if p == 3
+        ξ = rand(rng)
+        w = 1.0 + (log(ξ + (1.0 - ξ)*exp(-2κ))/κ)
+        return w::Float64
+    end
 
     r = (p - 1) / 2.0
     betad = Beta(r, r)
@@ -76,47 +84,8 @@ end
 _vmf_genw(rng::AbstractRNG, s::VonMisesFisherSampler) =
     _vmf_genw(rng, s.p, s.b, s.x0, s.c, s.κ)
 
-function _vmf_rotmat(u::Vector{Float64})
-    # construct a rotation matrix Q
-    # s.t. Q * [1,0,...,0]^T --> u
-    #
-    # Strategy: construct a full-rank matrix
-    # with first column being u, and then
-    # perform QR factorization
-    #
-
-    p = length(u)
-    A = zeros(p, p)
-    copyto!(view(A,:,1), u)
-
-    # let k the be index of entry with max abs
-    k = 1
-    a = abs(u[1])
-    for i = 2:p
-        @inbounds ai = abs(u[i])
-        if ai > a
-            k = i
-            a = ai
-        end
-    end
-
-    # other columns of A will be filled with
-    # indicator vectors, except the one
-    # that activates the k-th entry
-    i = 1
-    for j = 2:p
-        if i == k
-            i += 1
-        end
-        A[i, j] = 1.0
-    end
-
-    # perform QR factorization
-    Q = Matrix(qr!(A).Q)
-    if dot(view(Q,:,1), u) < 0.0  # the first column was negated
-        for i = 1:p
-            @inbounds Q[i,1] = -Q[i,1]
-        end
-    end
-    return Q
+function _vmf_householder_vec(u::Vector{Float64})
+    u = normalize(u)
+    u[1] -= 1.0
+    return normalize!(u)
 end