--- a/test/vonmisesfisher.jl
+++ b/test/vonmisesfisher.jl
@@ -22,6 +22,35 @@ function gen_vmf_tdata(n::Int, p::Int,
     return X
 end
 
+
+function test_genw3(κ::Real, ns::Int, rng::Union{AbstractRNG, Missing} = missing)
+    p = 3
+
+    if ismissing(rng)
+        μ = randn(p)
+    else
+        μ = randn(rng, p)
+    end
+    μ = μ ./ norm(μ)
+
+    s = Distributions.VonMisesFisherSampler(μ, float(κ))
+
+    genw3_res = [Distributions._vmf_genw3(rng, s.p, s.b, s.x0, s.c, s.κ) for _ in 1:ns]
+    genwp_res = [Distributions._vmf_genwp(rng, s.p, s.b, s.x0, s.c, s.κ) for _ in 1:ns]
+
+    @test isapprox(mean(genw3_res), mean(genwp_res), atol=0.01)
+    @test isapprox(std(genw3_res), std(genwp_res), atol=0.01/κ)
+
+    # test mean and stdev against analytical formulas
+    coth_κ = coth(κ)
+    mean_w = coth_κ - 1/κ
+    var_w = 1 - coth_κ^2 + 1/κ^2
+
+    @test isapprox(mean(genw3_res), mean_w, atol=0.01)
+    @test isapprox(std(genw3_res), sqrt(var_w), atol=0.01/κ)
+end
+
+
 function test_vonmisesfisher(p::Int, κ::Real, n::Int, ns::Int,
                              rng::Union{AbstractRNG, Missing} = missing)
     if ismissing(rng)
@@ -65,6 +94,7 @@ function test_vonmisesfisher(p::Int, κ::Real, n::Int, ns::Int,
         x = rand(rng, d)
     end
     @test norm(x) ≈ 1.0
+    @test insupport(d, x)
 
     if ismissing(rng)
         X = rand(d, n)
@@ -73,6 +103,7 @@ function test_vonmisesfisher(p::Int, κ::Real, n::Int, ns::Int,
     end
     for i = 1:n
         @test norm(X[:,i]) ≈ 1.0
+        @test insupport(d, X[:,i])
     end
 
     # MLE
@@ -118,5 +149,11 @@ ns = 10^6
                                                                            (5, 2.0),
                                                                            (2, 2)]
         test_vonmisesfisher(p, κ, n, ns, rng)
+    end
+
+    if !ismissing(rng)
+        @testset "Testing genw with $key at (3, $κ)" for κ in [0.1, 0.5, 1.0, 2.0, 5.0]
+            test_genw3(κ, ns, rng)
+        end
     end
 end