Resolves #1161 

Had some free time so I just implemented it right away lol