@emerali 2020-09-03
# Benchmarks:

## master:
### p = 3
```
julia> vmf = VonMisesFisher(rand(3));

julia> @benchmark rand($vmf)
BenchmarkTools.Trial: 
  memory estimate:  1.22 KiB
  allocs estimate:  10
  --------------
  minimum time:     3.841 μs (0.00% GC)
  median time:      4.130 μs (0.00% GC)
  mean time:        4.639 μs (0.65% GC)
  maximum time:     157.598 μs (95.62% GC)
  --------------
  samples:          10000
  evals/sample:     8

julia> @benchmark rand($vmf, 1000)
BenchmarkTools.Trial: 
  memory estimate:  24.63 KiB
  allocs estimate:  11
  --------------
  minimum time:     319.272 μs (0.00% GC)
  median time:      332.700 μs (0.00% GC)
  mean time:        340.943 μs (0.15% GC)
  maximum time:     1.573 ms (73.92% GC)
  --------------
  samples:          10000
  evals/sample:     1

```

### p = 1000
```
julia> vmf = VonMisesFisher(rand(1000));

julia> @benchmark rand($vmf)
BenchmarkTools.Trial: 
  memory estimate:  31.08 MiB
  allocs estimate:  16
  --------------
  minimum time:     102.893 ms (0.00% GC)
  median time:      121.886 ms (0.63% GC)
  mean time:        162.252 ms (5.46% GC)
  maximum time:     377.688 ms (0.67% GC)
  --------------
  samples:          31
  evals/sample:     1

julia> @benchmark rand($vmf, 1000)
BenchmarkTools.Trial: 
  memory estimate:  38.70 MiB
  allocs estimate:  17
  --------------
  minimum time:     449.606 ms (0.27% GC)
  median time:      675.868 ms (0.15% GC)
  mean time:        657.465 ms (4.89% GC)
  maximum time:     839.365 ms (0.09% GC)
  --------------
  samples:          8
  evals/sample:     1

```

## this PR

### p = 3
```
julia> vmf = VonMisesFisher(rand(3));

julia> @benchmark rand($vmf)
BenchmarkTools.Trial: 
  memory estimate:  336 bytes
  allocs estimate:  3
  --------------
  minimum time:     242.661 ns (0.00% GC)
  median time:      256.136 ns (0.00% GC)
  mean time:        282.311 ns (1.80% GC)
  maximum time:     2.801 μs (82.71% GC)
  --------------
  samples:          10000
  evals/sample:     413

julia> @benchmark rand($vmf, 1000)
BenchmarkTools.Trial: 
  memory estimate:  23.73 KiB
  allocs estimate:  4
  --------------
  minimum time:     114.304 μs (0.00% GC)
  median time:      116.085 μs (0.00% GC)
  mean time:        123.298 μs (0.32% GC)
  maximum time:     1.100 ms (81.23% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

### p = 1000
```
julia> vmf = VonMisesFisher(rand(1000));

julia> @benchmark rand($vmf)
BenchmarkTools.Trial: 
  memory estimate:  23.81 KiB
  allocs estimate:  3
  --------------
  minimum time:     13.781 μs (0.00% GC)
  median time:      14.682 μs (0.00% GC)
  mean time:        15.868 μs (3.30% GC)
  maximum time:     1.195 ms (92.74% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark rand($vmf, 1000)
BenchmarkTools.Trial: 
  memory estimate:  7.64 MiB
  allocs estimate:  4
  --------------
  minimum time:     10.918 ms (0.00% GC)
  median time:      11.292 ms (0.00% GC)
  mean time:        12.387 ms (3.96% GC)
  maximum time:     83.618 ms (83.11% GC)
  --------------
  samples:          404
  evals/sample:     1

```