@johnczito 2020-10-09
Thanks for this! Looks good to me. My only suggestion is to take this opportunity to make the unit tests more convincing. Might include:

- Test the two `w` samplers against one another in the `p = 3` case;
- Check that the `x` that's constructed at the end of `rand!` is the same as if you naively computed `H(μ - e₁ ) * t`.

I am in favor of things collapsing to the special cases when `p = 1` or `κ = 0` (especially the latter), but I think that should be handled in a separate PR.